HomeView, WriteProgress+ViewModel changes Изменения в соответствии с комментариями сделал поля  writeProgressOffset, blurRadius private В HomeView WriteProgress-ViewModel переименовал в WriteProgress+ViewModel и сделал класс ViewModel final изменил метод saveReadingSession Избавился от форс анврапа в getSurahName и добавил обработку ошибок FetchingDataError checkNumbers -> private checkNumbers
